requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
dj-static==0.0.6
Django==5.2
django-stdimage==6.0.2
fonttools==4.57.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-decouple==3.8
sqlparse==0.5.3
static3==0.7.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
whitenoise==6.9.0
dj_database_url==2.1.0
python-dotenv~=1.1.0
cloudinary==1.44.0
django-cloudinary-storage==0.3.0
